requirements.txt
amqp==5.2.0
APScheduler==3.6.3
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
backcall==0.2.0
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Deprecated==1.2.14
dj-database-url==1.0.0
Django==5.0.4
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-timezone-field==6.1.0
executing==2.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.7
ipython==8.23.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.18.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.2.0
wrapt==1.16.0
